Task 15 - 20 completed for check

+++ Classroom/10.09.2022/Task10.java
@@ -1,12 +1,12 @@
 public class Task10 {
 	public static void main(String[] args) {
-		int a = Integer.parseInt(args[0]);
+		long a = Long.parseLong(args[0]);
 
 		if (a < 0) {
 			a *= -1;
 		}
 
-		int sum = 0;
+		long sum = 0;
 		while(a > 0) {
 			sum += a % 10;
 			a /= 10;
